refactor(config): extract formatting helpers to _formatting.py
Convert config.py into a config/ package and extract 7 formatting
functions (268 lines) into _formatting.py. __init__.py drops from
1,182 to ~910 lines. Pure function extraction — only read config,
no mutation.

Closes #374
